+++ tachyon/zk/air/plonky3/base/two_adic_multiplicative_coset.h
@@ -26,7 +26,7 @@ class TwoAdicMultiplicativeCoset {
  public:
   constexpr TwoAdicMultiplicativeCoset() = default;
 
-  TwoAdicMultiplicativeCoset(size_t log_n, const F& shift) {
+  TwoAdicMultiplicativeCoset(uint32_t log_n, const F& shift) {
     domain_.reset(static_cast<math::Radix2EvaluationDomain<F>*>(
         math::Radix2EvaluationDomain<F>::Create(size_t{1} << log_n)
             ->GetCoset(shift)
@@ -88,7 +88,7 @@ class TwoAdicMultiplicativeCoset {
     CHECK_EQ(domain_->offset(), F::One());
     CHECK_NE(coset_shift, F::One());
     CHECK_GE(domain_->log_size_of_group(), coset.domain()->log_size_of_group());
-    size_t rate_bits =
+    uint32_t rate_bits =
         coset.domain()->log_size_of_group() - domain_->log_size_of_group();
     F s_pow_n = coset_shift.ExpPowOfTwo(domain_->log_size_of_group());
